Removed checksum for incorporated result, moved mock entity
