+++ model/flow/incorporated_result.go
@@ -21,15 +21,8 @@ func NewIncorporatedResult(incorporatedBlockID Identifier, result *ExecutionResu
 	}
 }
 
-// ID implements flow.Entity.ID for IncorporatedResult to make it capable of
-// being stored directly in mempools and storage.
+// ID returns unique identifier for the IncorporatedResult struct.
 func (ir *IncorporatedResult) ID() Identifier {
-	return MakeID([2]Identifier{ir.IncorporatedBlockID, ir.Result.ID()})
-}
-
-// CheckSum implements flow.Entity.CheckSum for IncorporatedResult to make it
-// capable of being stored directly in mempools and storage.
-func (ir *IncorporatedResult) Checksum() Identifier {
 	return MakeID(ir)
 }
